support toggle buttons in the plugins
A toggle is a button that has 2 images and a "selected" state. When you
tap the button the image show change to the "selected" image and be the
default state until the button state changes. This is useful to, for
example set the current Scene in OBS and let the user know in the deck
wich scene is selected